requirements.txt
paho-mqtt
picamera2